refac: add more independent component
