+++ src/App.css
@@ -46,3 +46,4 @@
   display: flex;
   flex-direction: column;
 }
+
