+++ src/app/WelcomeScreen/WelcomeScreen.jsx
@@ -1,11 +1,9 @@
 import { NumberInput } from "../../components/NumberInput/NumberInput"
-import { ButtonsGroup } from "./ButtonsGroup";
-import { InputsGroup } from "./InputsGroup";
+import { ButtonsGroup } from "../../components/ButtonsGroup/ButtonsGroup";
+import { InputsGroup } from "../../components/InputsGroup/InputsGroup";
 
 
 export const WelcomeScreen = () => {
-
-
     return (
         <div className="screen">
             <NumberInput />
