+++ src/components/ButtonsGroup/ButtonsGroup.jsx
@@ -1,0 +1,15 @@
+import { Button } from "../Button/Button"
+import buttons from '../buttonsData'
+
+export const ButtonsGroup = () => {
+    return (
+        <>
+            {Object.values(buttons).map((item, index) => (
+                <Button key={index} {...item} />
+            ))}
+        </>
+    )
+
+
+}
+
