+++ src/components/InputsGroup/InputsGroup.jsx
@@ -1,0 +1,14 @@
+import { SelectInput } from "../SelectInput/SelectInput"
+import inputs from "../inputsData"
+
+export const InputsGroup = () => {
+    return (
+        <>
+            {Object.values(inputs).map((input, index) => (
+                <SelectInput key={index} {...input} />
+            ))}
+        </>
+    )
+
+
+}
