+++ src/components/buttonsData.js
@@ -1,0 +1,16 @@
+const buttons = {
+    startNewGameBtn: {
+        text: 'Start new game',
+        className: 'new-game-btn',
+        buttonType: 'submit',
+        formaction: ''
+    },
+    seeStatisticsBtn: {
+        text: 'See my statistics',
+        className: 'see-stat-btn',
+        buttonType: 'button',
+        formaction: ''
+    }
+};    
+
+export default buttons
