+++ src/components/inputsData.js
@@ -1,0 +1,30 @@
+const inputs = {
+    categoryInput: {
+        text: 'Category:',
+        id: 'category',
+        options: ['JS', 'React'],
+        className: 'category-input'
+    },
+    difficultyInput: {
+        text: 'Choose difficulty',
+        id: 'difficulty',
+        options: ['easy', 'medium', 'hard'],
+        className: 'difficulty-input'
+
+    },
+    typeInput: {
+        text: 'Choose type',
+        id: 'type',
+        options: ['true', 'false'],
+        className: 'type-input'
+
+    },
+    timeInput: {
+        text: 'Choose time',
+        id: 'time',
+        options: ['1m', '2m', '5m'],
+        className: 'time-input'
+    },
+};
+
+export default inputs
